Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit 1455b66aaebb27cb87755c6ca86b0643f7c76017, reversing
changes made to d1d5adc8c07059eb1fd0e578f4ae0e16fc4a9576.

+++ src/lib/EEPROM/elrs_eeprom.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <stdint.h>
-#include <cstddef>
 
 #define RESERVED_EEPROM_SIZE 1024
 
@@ -18,14 +17,14 @@ public:
     template <typename T> void Get(uint32_t addr, T &value)
     {
         uint8_t* p = (uint8_t*)(void*)&value;
-        size_t   i = sizeof(value);
+        uint8_t  i = sizeof(value);
         while(i--)  *p++ = ReadByte(addr++);
     };
 
     template <typename T> const void Put(uint32_t addr, const T &value)
     {
         const uint8_t* p = (const uint8_t*)(const void*)&value;
-        size_t         i = sizeof(value);
+        uint8_t        i = sizeof(value);
         while(i--)  WriteByte(addr++, *p++);
     };
 };
